+++ apps/web/src/lib/components/forms/TemplateBasics.svelte
@@ -48,5 +48,5 @@
 
 <label class={labelClass}>
   <span class={spanClass}>Image *</span>
-  <input type="text" bind:value={image} placeholder="tsukisama9292/ow-kasmvnc-ubuntu-dini:jammy" class={inputClass} />
+  <input type="text" bind:value={image} placeholder="tsukisama9292/ow-kasmvnc-ubuntu:jammy" class={inputClass} />
 </label>